fix: user&product action files
