+++ components/settings/LocationChooser.tsx
@@ -12,7 +12,7 @@ import {
   FormLabel,
   FormMessage,
 } from '@/components/ui/form'
-import { saveUserLocation } from '@/lib/action'
+import { saveUserLocation } from '@/lib/useraction'
 
 const formSchema = z.object({
   city: z
