+++ models/product-model.ts
@@ -7,17 +7,3 @@ export type ProductType = {
   location: string;
   status: string;
 };
-
-export type FullProductType = {
-  id: string;
-  title: string;
-  description: string ;
-  price: number;
-  currency: string;
-  quantity: number;
-  location: string ;
-  status: string;
-  sellerId: string;
-  image: string ;
-  createdAt: Date;
-};
